added smbus code to folder. Removed interrups on ADC. adc_sample now returns conversions from all 8 pins in one go. code compiled and tested on slug.

+++ firmware/dio-adc/trunk/adc.h
@@ -15,9 +15,9 @@
     Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA. */
 #ifndef __ADC_H
 #define __ADC_H
-
+#include <stdint.h>
 void adc_init ( void );
-int * adc_sample ( void );
+uint16_t*  adc_sample(void);
 
 
 #endif 	/* __ADC_H  */
